OpenGl project
Man on the Moon application demonstrates the technological basis for advanced space exploration, enabling the first robotic and human missions to the Moon. A visualization of the planet Earth from the moon is depicted along with other celestial bodies. It includes user interaction through keyboard events for erecting of the flag and for placing a sapling on the lunar surface.